+++ packages/code-du-travail-frontend/src/outils/CommonSteps/Informations/index.tsx
@@ -2,13 +2,14 @@ import React from "react";
 
 import { PubliQuestion } from "./components";
 import { MatomoActionEvent } from "../../../lib";
-import { PublicodesInformation } from "./store";
+import { CommonInformationsStoreError, PublicodesInformation } from "./store";
 import { RuleType } from "@socialgouv/modeles-social";
+import { InlineError } from "../../common/ErrorField";
 
 export type InformationStepProps = {
   onChange: (key: string, value: unknown, type: RuleType | undefined) => void;
   informations: PublicodesInformation[];
-  errors: Record<string, string>;
+  errors: CommonInformationsStoreError;
 };
 
 const CommonInformationStep = ({
@@ -17,7 +18,7 @@ const CommonInformationStep = ({
   errors,
 }: InformationStepProps): JSX.Element => (
   <>
-    {informations.map((info) => {
+    {informations.map((info, index) => {
       return (
         <PubliQuestion
           key={info.id}
@@ -28,10 +29,14 @@ const CommonInformationStep = ({
           onChange={(v) =>
             onChange(info.question.rule.nom, v, info.question.rule.cdtn?.type)
           }
-          error={errors[info.question.rule.nom] ?? undefined}
+          error={errors.errorInformations[info.question.rule.nom] ?? undefined}
+          autoFocus={index === 0}
         />
       );
     })}
+    {errors.errorPublicodes && (
+      <InlineError>{errors.errorPublicodes}</InlineError>
+    )}
   </>
 );