+++ packages/code-du-travail-frontend/src/outils/common/Agreement/EnterpriseSearch/SelectedEnterprise.tsx
@@ -25,6 +25,7 @@ const SelectedEnterprise = ({
           event.preventDefault();
           onRemoveEnterprise();
         }}
+        tabIndex={1}
       >
         {enterprise.simpleLabel}
       </Selected>