+++ packages/code-du-travail-frontend/src/outils/publicodes/index.ts
@@ -1,4 +1,2 @@
-export * from "./context";
-export * from "./helpers";
-export * from "./usePublicodes";
-export * from "./utils";
+export * from "./common";
+export * from "./indemnite-licenciement";