+++ packages/code-du-travail-modeles/src/publicodes/types.ts
@@ -44,6 +44,7 @@ export interface Rule extends PubliRule {
 
 export type NotificationAffichage =
   | "conventionnel"
+  | "default"
   | "légal et conventionnel"
   | "légal";